requirements.txt
functions-framework==3.*
requests==2.32.*
streamlit
google-cloud-bigquery
pandas
requests
db-dtypes